change Time from int? to long?

+++ src/SiftScienceNet/Events/Order.cs
@@ -46,7 +46,7 @@ namespace SiftScienceNet.Events
         public bool? ExpeditedShipping { get; set; }
 
         [JsonProperty("$time")]
-        public int? Time { get; set; }
+        public long? Time { get; set; }
 
         [JsonProperty("$ip")]
         public string Ip { get; set; }
